imagersite/requirements.txt
apipkg==1.4
appnope==0.1.0
decorator==4.0.9
dj-database-url==0.4.1
Django==1.9.5
django-extensions==1.6.1
django-registration==2.0.4
django-widget-tweaks==1.4.1
djangorestframework==3.3.3
execnet==1.4.1
factory-boy==2.6.1
fake-factory==0.5.7
gnureadline==6.3.3
ipaddress==1.0.16
ipython==4.1.2
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
Pillow==3.2.0
psycopg2==2.6.1
ptyprocess==0.5.1
py==1.4.31
pytest==2.9.1
pytest-xdist==1.14
python-dateutil==2.5.2
redis==2.10.5
simplegeneric==0.8.1
six==1.10.0
sorl-thumbnail==12.3
traitlets==4.2.1
